requirements.txt
langchain-openai
langchain
langchain-community
openai
tiktoken
chromadb
sentence-transformers
fastapi
uvicorn[standard]
pydantic==2.*
python-dotenv
unstructured
pypdf
beautifulsoup4
lxml
markdown
numpy
scikit-learn
pandas
streamlit
requests
